engram/core/memory/working/projects/agent-harness/IN/harness/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "harness"
version = "0.1.0"
description = "Agent harness CLI and tooling"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.40",
    "openai>=1.40.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
test = ["pytest>=8"]
engram = ["agent-memory-mcp[search]"]

[project.scripts]
harness-report = "harness.report:main"

[tool.setuptools]
package-dir = { harness = "." }
packages = [
    "harness",
    "harness.modes",
    "harness.tools",
    "harness.tools.fs",
    "harness.tools.search",
]
include-package-data = true

[tool.setuptools.package-data]
harness = ["pricing.json"]

[tool.pytest.ini_options]
testpaths = ["tests"]
